refactor(llm): split model.rs (1349 lines) into model/ modules
Pure mechanical split, byte-identical logic, public API unchanged
(hermes_llm::Transformer, model::{Transformer,InferenceState,...} still resolve):
  model/{norm,rope,attention,ffn,mamba,block,transformer}.rs + mod.rs re-exports.
Only fields crossing new module boundaries widened to pub(crate). 26 tests pass,
clippy -D warnings clean on default + metal,remote.
